fix(ci): MSVC C4365 signed/unsigned 불일치 수정 (헤더 포함)
- xpe_calib_mode.cpp: static_assert에서 == 6 → == 6u
- xpe_defect_gen.cpp: uint8_t | uint8_t 결과에 static_cast<uint8_t> 추가
- xpe_calib_generate_gain.cpp: y_vals[sNumLevels - 1] → sNumLevels - size_t{1}
- xpe_preprocess_internal.h: sizeof(CalibFileHeader) == 64 → == 64u
- xcal_format.h: sizeof/offsetof static_assert 정수 리터럴 모두 u 접미사 추가

+++ modules/preprocess/src/xpe_defect_gen.cpp
@@ -362,7 +362,7 @@ void merge_bpm(const std::vector<uint8_t>& dark_bpm,
 
     for (size_t i = 0; i < num_pixels; ++i) {
         // Merge using logical OR: 0, 1, 2, or 3 (both)
-        bpm_out[i] = dark_bpm[i] | bright_bpm[i];
+        bpm_out[i] = static_cast<uint8_t>(dark_bpm[i] | bright_bpm[i]);
     }
 }
